requirements.txt
Flask~=3.0.3
gspread~=6.1.4
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests~=2.32.3
protobuf~=5.28.3
Flask-Cors~=5.0.0